+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### New features
+
+- Add new `Sidekiq/UnsafeQueueOperation` cop.
+- Add new `SidekiqPro/PreferDeleteJob` cop.
+
 ## 0.1.0 - 2026-01-20
 
 - Initial release.